[Refactor] 공정 분석 화면 구조 정리
